requirements.txt
streamlit
Pillow
tensorflow

